backend/requirements.txt
# backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.7
drf-yasg==1.21.5
Pillow==10.0.1
cryptography==41.0.7
geopy==2.3.0
requests==2.31.0
django-filter==23.3
